responsiveness edits
did not end up using styled components however finished responsiveness on product cards and directory preview
