+++ src/components/category-preview/category-preview.styles.scss
@@ -4,8 +4,8 @@
   margin-bottom: 90px;
 
   .title {
-    font-size: 28px;
     margin-bottom: 25px;
+    font-family: 'Cormorant Garamond';
     cursor: pointer;
   }
