+++ src/components/directory-preview/directory-preview.styles.scss
@@ -3,8 +3,7 @@
   flex-direction: column;
   justify-content: center;
   align-items: stretch;
-  height: 100vh;
-  margin: 50px 20px;
+  margin: 20px 20px; /* Adjust the top margin to your preference */
   padding: 20px 60px;
 
   .directory-title {
@@ -17,51 +16,44 @@
 
   .directory-preview {
     display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+    grid-template-columns: repeat(4, 1fr);
     column-gap: 20px;
   }
 }
 
+// Media Queries
 /* Extra small devices (phones, 600px and down) */
 @media only screen and (max-width: 600px) {
-  .directory-preview {
-    // grid-template-columns: repeat(1, 1fr);
-    // column-gap: 40px;
-    background-color: purple;
+  .directory-preview-container {
+    .directory-preview {
+      grid-template-columns: repeat(1, 1fr);
+    }
   }
 }
 
 /* Small devices (portrait tablets and large phones, 600px and up) */
 @media only screen and (min-width: 600px) {
-  .directory-preview {
-    // grid-template-columns: repeat(1, 1fr);
-    // column-gap: 10px;
-    background-color: pink;
+  .directory-preview-container {
+    .directory-preview {
+      grid-template-columns: repeat(2, 1fr);
+    }
   }
 }
 
 /* Medium devices (landscape tablets, 768px and up) */
 @media only screen and (min-width: 768px) {
-  .directory-preview {
-    // display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    // column-gap: 20px;
-    background-color: green;
+  .directory-preview-container {
+    .directory-preview {
+      grid-template-columns: repeat(2, 1fr);
+    }
   }
 }
 
-// Medium devices (landscape tablets and medium desktops, 992px and up)
+/*Medium devices (landscape tablets and medium desktops, 992px and up)*/
 @media only screen and (min-width: 992px) {
-  .directory-preview {
-    grid-template-columns: repeat(2, 1fr);
-    background-color: orange;
+  .directory-preview-container {
+    .directory-preview {
+      grid-template-columns: repeat(4, 1fr);
+    }
   }
-}
-
-/* Extra large devices (large laptops and desktops, 1400px and up) */
-@media only screen and (min-width: 1400px) {
-}
-
-// Extra-large devices (1600px and up)
-@media only screen and (min-width: 1600px) {
 }
