+++ src/components/product-card/product-card.styles.scss
@@ -1,7 +1,7 @@
-$desktop: 1200px;
-$laptop: 992px;
-$tablet: 768px;
-$phone: 480px;
+// $desktop: 1200px;
+// $laptop: 992px;
+// $tablet: 768px;
+// $phone: 480px;
 
 .product-card-container {
   width: 100%;
@@ -30,7 +30,6 @@ $phone: 480px;
     top: 255px;
     display: none;
   }
-
   &:hover {
     img {
       opacity: 0.8;
@@ -60,64 +59,45 @@ $phone: 480px;
     }
   }
 }
-//desktop 1200 tablet 992
-@media (min-width: $desktop) and (min-width: $tablet) {
-  .product-img {
-    height: 450px;
-    width: 100%;
-
-    img {
-      min-height: 100%;
-      object-fit: cover;
-    }
-  }
-}
-// @media (min-width: $laptop) {
-// }
-// @media (min-width: $tablet) {
-// }
-// @media (min-width: $phone) {
-// }
-
-// /* Extra small devices (phones, 600px and down) */
-// @media only screen and (min-width: 600px) {
-// }
-// // Apply for screens up to 600px
-// @media only screen and (max-width: 600px) {
-// }
 
 // Apply for screens up to 768px THIS IS FOR A TABLET
 @media only screen and (max-width: 768px) {
 }
 // Apply for screens up to 992px LAPTOP
 @media only screen and (max-width: 992px) {
-  .product-img {
-    height: 407px;
-    width: 100%;
-
-    img {
+  .product-card-container {
+    height: 400px;
+    .product-img {
+      height: 350px;
       width: 100%;
-      height: auto;
-      object-fit: cover;
-      border-radius: 4px;
+
+      img {
+        width: 100%;
+        height: 100%;
+        object-fit: cover;
+        border-radius: 4px;
+      }
     }
   }
 }
 
 // Apply for screens up to 1200px
-@media only screen and (max-width: 1400px) {
-  .product-img {
-    height: 407px;
-    width: 100%;
+// @media only screen and (max-width: 1400px) {
+//   .product-card-container {
+//     height: 500px;
+//     .product-img {
+//       height: 350px;
+//       width: 100%;
 
-    img {
-      width: 100%;
-      height: auto;
-      object-fit: cover;
-      border-radius: 4px;
-    }
-  }
-}
+//       img {
+//         width: 100%;
+//         height: 100%;
+//         object-fit: cover;
+//         border-radius: 4px;
+//       }
+//     }
+//   }
+// }
 
 // Apply for screens up to 1600px
 @media only screen and (max-width: 1600px) {
@@ -130,23 +110,6 @@ $phone: 480px;
       img {
         width: 100%;
         min-height: 100%;
-      }
-    }
-  }
-}
-
-@media only screen and (max-width: 1800px) and (min-width: 1600px) {
-  .product-card-container {
-    height: 600px;
-
-    .product-img {
-      height: 450px;
-      width: 100%;
-
-      img {
-        width: auto;
-        min-height: 100%;
-        object-fit: cover;
       }
     }
   }
