+++ src/routes/categories-preview/categories-preview.jsx
