+++ src/utils/firebase/shoptTitle.utils.jsx
@@ -1,0 +1,5 @@
+export const shopTitle = (title) => {
+  return title
+    .replace(/_/g, ' ')
+    .replace(/\b\w/g, (firstLetter) => firstLetter.toUpperCase())
+}
